refactor(wallet-app): improve utils imports
