+++ 0-1/projects/basic-wallet-app/server/middlewares/error.middlewares.js
@@ -1,6 +1,6 @@
 import mongoose from 'mongoose';
 
-import {ApiError} from '../utils/ApiError.js';
+import {ApiError} from '../utils/index.js';
 
 /**
  *
